--- autorec_ai/ingest/config.py
@@ -1,6 +1,0 @@
-RAW_DATA_PATH = "data/raw"
-METADATA_PATH = "data/car_metadata"
-# todo: i need to do this due to python and pytorch shenanigans on my machine. will fix it
-GROUPED_BY_CAR_MAKE_MODEL_YEAR_DATA_PATH = "../../data/raw_grouped_by_car_make_model_year"
-PROCESSED_DATA_PATH = "../../data/processed/augmented_grouped_by_car_make_model_year"
-PROCESSED_AUGMENTATION_ANALYSIS_REPORT_DATA_PATH = "../../data/processed"
