+++ autorec_ai/utils/config.py
@@ -1,0 +1,5 @@
+RAW_DATA_PATH = "data/raw"
+METADATA_PATH = "data/car_metadata"
+GROUPED_BY_CAR_MAKE_MODEL_YEAR_DATA_PATH = "data/raw_grouped_by_car_make_model_year"
+PROCESSED_DATA_PATH = "data/processed/augmented_grouped_by_car_make_model_year"
+PROCESSED_AUGMENTATION_ANALYSIS_REPORT_DATA_PATH = "data/processed"
